third_party/blink/web_tests/platform/generic/fast/forms/radio/radio-indeterminate-pseudo-class-expected.txt

This tests :indeterminate pseudo class matches when there is no checked radio button in a radio button group.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

If a name isn't specified, it is in its own radio group.
PASS r1_1.matches(":indeterminate") is true
Setting indeterminate property shouldn't affect radio buttons.
PASS r1_1.matches(":indeterminate") is true
PASS r1_1.matches(":indeterminate") is false
PASS r1_1.matches(":indeterminate") is false
Changing checked should affect the matching of other buttons within the radio group.
PASS r2_1.matches(":indeterminate") is true
PASS r2_2.matches(":indeterminate") is true
PASS r2_3.matches(":indeterminate") is true
PASS r2_1.matches(":indeterminate") is false
PASS r2_2.matches(":indeterminate") is false
PASS r2_3.matches(":indeterminate") is false
PASS r2_1.matches(":indeterminate") is false
PASS r2_2.matches(":indeterminate") is false
PASS r2_3.matches(":indeterminate") is false
PASS r2_1.matches(":indeterminate") is true
PASS r2_2.matches(":indeterminate") is true
PASS r2_3.matches(":indeterminate") is true
Adding/removing a button from a group should affect the matching.
PASS r3_1.matches(":indeterminate") is false
PASS r4_1.matches(":indeterminate") is true
PASS r4_2.matches(":indeterminate") is true
PASS r3_1.matches(":indeterminate") is false
PASS r4_1.matches(":indeterminate") is false
PASS r4_2.matches(":indeterminate") is false
PASS r3_1.matches(":indeterminate") is false
PASS r4_1.matches(":indeterminate") is true
PASS r4_2.matches(":indeterminate") is true
PASS successfullyParsed is true

TEST COMPLETE

      
